backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymongo==4.11.1
redis==5.2.1
python-dotenv==1.0.1
razorpay==1.4.2
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic==2.10.6
email-validator==2.2.0
openai==1.61.1
pytest==8.3.4
